LayoutTests/fast/canvas/philip/tests/2d.imageData.object.round.html

<!DOCTYPE html>
<title>Canvas test: 2d.imageData.object.round</title>
<script src="../tests.js"></script>
<link rel="stylesheet" href="../tests.css">
<link rel="prev" href="2d.imageData.object.wrap.html" title="2d.imageData.object.wrap">
<link rel="next" href="2d.imageData.put.null.html" title="2d.imageData.put.null">
<body class="show_output">
<p>
 <a href="2d.imageData.object.wrap.html" accesskey="p" title="[p] 2d.imageData.object.wrap">&lt;</a>
 <a href="index.html">[index]</a>
 <a href="2d.imageData.put.null.html" accesskey="n" title="[n] 2d.imageData.put.null">&gt;</a>
<h1><a href="index.2d.html">2d</a>.<a href="index.2d.imageData.html">imageData</a>.<a href="index.2d.imageData.object.html">object</a>.round</h1>
<p class="desc">ImageData.data rounds numbers with round-to-zero</p>
<div class="refs">References:
<ul>
<li><a href="spec.html#testrefs.2d.pixelarray.modify">2d.pixelarray.modify</a>

</ul>
</div>

<p class="output">Actual output:</p>
<canvas id="c" class="output" width="100" height="50"><p class="fallback">FAIL (fallback content)</p></canvas>

<ul id="d"></ul>
<script>
_addTest(function(canvas, ctx) {

var imgdata = ctx.getImageData(0, 0, 10, 10);
imgdata.data[0] = 0.499;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
imgdata.data[0] = 0.5;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
imgdata.data[0] = 0.501;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
imgdata.data[0] = 1.499;
_assertSame(imgdata.data[0], 1, "imgdata.data[\""+(0)+"\"]", "1");
imgdata.data[0] = 1.5;
_assertSame(imgdata.data[0], 2, "imgdata.data[\""+(0)+"\"]", "2");
imgdata.data[0] = 1.501;
_assertSame(imgdata.data[0], 2, "imgdata.data[\""+(0)+"\"]", "2");
imgdata.data[0] = 2.5;
_assertSame(imgdata.data[0], 2, "imgdata.data[\""+(0)+"\"]", "2");
imgdata.data[0] = 3.5;
_assertSame(imgdata.data[0], 4, "imgdata.data[\""+(0)+"\"]", "4");
imgdata.data[0] = 252.5;
_assertSame(imgdata.data[0], 252, "imgdata.data[\""+(0)+"\"]", "252");
imgdata.data[0] = 253.5;
_assertSame(imgdata.data[0], 253, "imgdata.data[\""+(0)+"\"]", "253");
imgdata.data[0] = 254.5;
_assertSame(imgdata.data[0], 254, "imgdata.data[\""+(0)+"\"]", "254");
imgdata.data[0] = 256.5;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
imgdata.data[0] = -0.5;
_assertSame(imgdata.data[0], 0, "imgdata.data[\""+(0)+"\"]", "0");
imgdata.data[0] = -1.5;
_assertSame(imgdata.data[0], 255, "imgdata.data[\""+(0)+"\"]", "255");


});
</script>
